CashFusion: Default fusion depth, if checked, to 3
Previously it would default to 1, which some members of the team think
is too low. So we set the default fusion depth to 3 when the checkbox is
first checked.  Note that the default behavior is for the checkbox to be
unchecked (fusions go on forever until paused). This change only affects
what happens when the user decides to enable the fusion depth limit
feature, in which case now it defauls to 3.