Fixup for Python 3.10
This commit contains some fixups for Python 3.10.  Apparently some of
the PyQt overloads disappeared that support int <-> float
interchangeably as args to C++ functions, which caused the app to not
work at all on Python 3.10.

Note: there are still extant bugs with using Python 3.10, but this
commit at least allows the app to (mostly) run ok.