Add new option: "Retire unused change addresses"  (#2408)
* Miscellaneous fixups to synchronizer.py and wallet.py

- Updated copyright in wallet.py and synchronizer.py
- Fixed a FIXME in synchronizer.py
- Fixed miscellanous python formatting/style issues found by PyCharm
- Added some typing annotations to make code more readable

* Refactor of synchronizer.py to support tracking of change addresses

This is an initial step in order to allow us to only subscribe to N
change addresses at a time.

* Fixed a comment that was too long

* Added mechanism to network.py to unsubscribe from addresses

New method in class Network: unsubscribe_from_scripthashes. If all
callbacks for a scripthash are gone, will send a
blockchain.scripthash.unsubscribe message to the server.

This will be used later to reap old change addresses.

* Unsubscribe from individual addresses on wallet close

This should keep the subs count down for unused addresses when wallets
are closed.  If the refct for a scripthash drops to 0, we send
blockchain.scripthash.unsubscribe to the server.

Note that an individual address may be "seen" or used by multiple
wallets so we only unsub when a scripthash's refct drops to 0 (meaning
no wallet cares about it anymore).

* Minor optimization in synchronizer.py subscribe_to_addresses

Don't build a list and then convert it to a dict. Build the dict
directly. This saves CPU.

* Get rid of checking if history is sorted properly

This check wastes CPU cycles and doesn't actually amount to any
functional change in Electron Cash.  Moreover, the check is not
guaranteed to work correctly in the case of mempool tx's appearing in
history, since the mempool txns have an unspecified order.

* Added tracking of change addresses and expiration candidates to synchronizer.py

These are the data structures we will use to expire change address
subscriptions.

* Implemented expiring old change address subs in synchronizer.py

If Synchronizer is constructed with positive, nonzero limit_change_subs,
then it will take an alternate code path to expire old change addresses.
It still runs through the entire change address list, and does briefly
subscribe to all change addresses to detect changes, however it then
unsubscribes when limit_change_subs is exceeded for change addresses.

* Added ability to enable/disable retiring of change in QT GUI

Default is enabled with a threshold of older than: 1000

* Refactor: Precede "private" methods in synchronizer.py with '_'

* Added periodid check that we are not over chagne addr subs limit

* Small nit/optimization

Create the huge int once for the sorter in _check_change_subs_limits
(should save CPU cycles).

* Added custom messaging when changing change retire settings

Tell user that they need to close and reopen the wallet (rather than all
of electron cash), in order for settings to take effect.

* Optimization: Persist the retired change addresses to wallet storage

This allows us to have very fast startup times in the happy case where
the retired change addresses have no new history.

Note that now it is possible for the wallet to not realize if it was
sent funds to a "retired" change address.  Such is life.  User can
disable the feature to revert to old behavior or they can select
"rebuild wallet history" in that corner case.

* Update address tab GUI code to gray out "retired" change addresses

This at least provides the user with some visual indicator of what is
happening.

* Only load the retired change set if using the limit_change option

Otherwise don't bother loading the set as it may cause trouble.

* Fixed undefined variable bug