Add LNS support (#2414)
* Add LNS support

* Review fixes

* Add LNS config to settings tab, use config in LNS lookups

* Add documentation to invalidate method

* Various fixups: Don't touch gui in main thread, fix unknown method error

- The VerifyingDialog was touching the GUI in the main thread in the
  error case.  This has been addressed.
- When receiving an error a non-existant method "wallet.show_error" was
  being called. We have just changed this to the existing method
  "wallet.print_error".

* Fix usage of Qt inside the `electroncash` module

Qt usage is not allowed in the `electroncash/` module. We removed all
assumptions that the app is running under Qt for the lns.py file, as
well as contacts.py, both of which life in `electroncash/`.

* Fix: Auto-remember "mine" LNS contacts

The app would always forget the "mine" LNS contacts and there was no way to
perma-save them. This has been addressed.

* Fix bug where editing LNS server crashes app

* Fix bug where send tab doesn't display LNS errors properly

Instead it was crashing with an exception reporter

* Respect user cashaddr / legacy UI settings for LNS address rendering

This wasn't being respected previously. Also fixed a potential bug where
we were touching wallet.contacts outside of the main thread.

* Make LNS contacts auto-refresh themselves every 10 mins

In the Qt app only.

Co-authored-by: Calin Culianu <calin.culianu@gmail.com>