Build: Install a MSVC linker wrapper to make PE timestamps deterministic
By default, the MSVC linker includes a timestamp in the PE header which
makes the build not reproducible. There is a `/Brepro` flag that makes
the timestamps deterministic.

Unfortunately setuptools provides no way to pass arguments to the MSVC
linker and there is no way to change the behavior using environment
variables. The `CL` environment variable also can not be used to pass
flags to the linker because setuptools calls the linker directly.

To fix this we compile and install a wrapper executable that adjusts the
arguments passed to the real linker binary to always include the
`/Brepro` flag.

Co-authored-by: Calin Culianu <calin.culianu@gmail.com>