CTOR workaround approach 2: Always display receives before sends (for the same block) (#2433)
* [qt] Add option to history tab to always display receives before sends

This option is default-on.  This implementation is an alternative to
PR #2431.  It is a second attempt to "undo the damage" of CTOR on the
wallet UI and avoid temporary negative intermediate balances.

The approach taken here is to always sort txns in the same block by
receives-first.  Thus all txns for the same block that have a net
positive effect on wallet balance will appear before these txns in the
same block with a net negative effect on wallet balance.  This is a
quick and performant way to avoid intermediate negative wallet balances
in the history tab (due to CTOR).

* Get rid of config UI for "Receives before sends" and just make always on

This was suggested in Telegram by a number of people and so we don't
even offer a config UI for disabling the "receives before sends" sorting
in the history list. It's just always enabled since it costs nothing and
the consensus was that it leads to less disconcerting behavior in the
UI.