Wine: Update to version 7.7
Wine version 7.6 improves the reliability of installing and using the
.NET framework, but was unusable to us because of issues with terminal
redirection. Wine version 7.7 fixed that issue, so we upgrade to that.

This also adds a wait for wineserver shutdown after installing the
linker wrapper as well as a `ngen executequeueditems` after installing
MSVC, which is currently disabled because of failures, but which
should be there to improve performance.