Qt: Disable destroyed_print_error unless a static flag is set
For performance, we now disable the "destroyed" log prints for -v mode
unless a static variable "debug_destroyed" is set to True. This is to
save CPU cycles as well as improve Qt stability.  It was observed that
on macOS and Windows sometimes the app would crash when the "destroyed"
signal is fired as the C++ object is deleted.  This is a bug in PyQt5.