Wine: Pass additional flags to reduce junk in the binaries
This passes the undocumented `/emittoolversioninfo:no` to stop the
linker from adding the "Rich" header which contains tool versions into
the file, see

https://developercommunity.visualstudio.com/t/Add-linker-option-to-strip-Rich-stamp/740443

This also adds the undocumented `/emitpogophaseinfo` to stop PGO
information to end up in the IMAGE_DEBUG_DIRECTORY.

Both of these seem to cause non-reproducible builds.

This also installs another wrapper for `cl.exe` which uses the post
arguments of the wrapper to pass these flags to the linker.